+++ src/web_manager.h
@@ -16,7 +16,6 @@ private:
 
     String buildStatusPage();
     String buildCalibrationPage(const String& message = "");
-    String buildRoadRatingPage();
     String buildUpdatePage();
     String buildUpdateResultPage(bool success);
     bool authenticateOTA();